update scirpts and hardhat config files

+++ scripts/deployUserContract.ts
@@ -5,6 +5,7 @@ import * as dotenv from "dotenv";
 import * as path from "path";
 import * as fs from "fs";
 import getRouter from "../config/getRouter";
+import { Peer } from "../utils/util";
 
 // Load environment variables from .env file
 dotenv.config({ path: path.join(__dirname, "../.env") });
@@ -48,14 +49,44 @@ async function main() {
   const contractAddress = await contract.getAddress();
   console.log(`${contractName} deployed to: ${contractAddress}`);
 
-  // Update equito.json with the new contract address
-  config.user_contract_address = contractAddress;
+  const chainId = hre.network.config.chainId;
+  if (!chainId) {
+    throw new Error(`Failed to fetch chainId!`);
+  } else {
+    console.log(`Received chainid: ${chainId}`);
+  }
+  // New peer to be added
+  const newPeer: Peer = {
+    chainId: chainId,
+    address: contractAddress,
+  };
 
-  // Write the updated config back to the file
-  fs.writeFileSync(configPath, JSON.stringify(config, null, 2));
-  console.log(
-    `Updated equito.json with new contract address: ${contractAddress}`,
+  // Check if the new peer already exists at current chainId
+  const existingPeerIndex = config.peers.findIndex(
+    (peer: Peer) => peer.chainId === newPeer.chainId,
   );
+
+  if (existingPeerIndex === -1) {
+    // If the peer does not exist, add the new peer
+    config.peers.push(newPeer);
+    // Write the updated configuration back to the file
+    fs.writeFileSync(configPath, JSON.stringify(config, null, 2));
+    console.log(
+      `Added new peer: ${JSON.stringify(newPeer)} into the config ${configPath}`,
+    );
+  } else {
+    const existingPeer = config.peers[existingPeerIndex];
+    if (existingPeer.address !== newPeer.address) {
+      // If the address does not match, update the address
+      config.peers[existingPeerIndex] = newPeer;
+      console.log(
+        `Updated peer at chainId ${newPeer.chainId} with new address: ${newPeer.address}`,
+      );
+    } else {
+      console.log(`Peer at chainId ${newPeer.chainId} already exists.`);
+    }
+  }
+
   process.exit(0);
 }
